Devices Rewrite: Fix Arena Rewards

+++ src/types.h
@@ -1945,7 +1945,7 @@ typedef struct
 {
     s16b r_idx; /* Monster (0 means victory prizing) */
     byte tval;  /* tval of prize (0 means no prize) */
-    byte sval;  /* sval of prize */
+    int  sval;  /* sval of prize, or effect_e for devices */
 } arena_type;
